requirements.txt
streamlit
supabase>=2.0.0
fpdf2
python-dotenv